Update 06_esp_pipeline.ipynb
Since I am only using patients with consecutive sessions in 12 weeks. The model only learns the pattern where there is no gap in between sessions. So when the action is 0, the model does not know what happens. This limitation causes an inflation of predicted score (continued upward trajectory) when no action is prescribed.